Implemented optional arguments.
Changed the procs which take an optional port to use optional
arguments instead of a final variable argument. Not only is this more
concise, but it avoids heap overflows by not constructing a list for
the variable argument. Now, the last test case could be reverted to be
an order-of-magnitude bigger.